+++ .npmignore
@@ -1,6 +1,6 @@
 coverage
 node_modules
-samples
+examples
 tests
 *.ts
 !*.d.ts
