+++ src/examples/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "samples",
+  "name": "examples",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
