+++ src/examples/tsconfig.json
